infra/cdk/pyproject.toml
[project]
name = "evidentia-cdk"
version = "0.1.0"
description = "CDK infrastructure for Evidentia"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "aws-cdk-lib>=2.160.0,<3.0.0",
  "constructs>=10.0.0,<11.0.0"
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["evidentia_cdk"]

